RR limit as integer

+++ VentilatorDaemon/Models/VentilatorSetting.cs
@@ -42,7 +42,7 @@ namespace VentilatorDaemon.Models
                 new VentilatorSetting<double>("FW", false, false), // firmware version
                 new VentilatorSetting<double>("FIO2", true, true), // oxygen level, 0.20 -> 1.0
                 new VentilatorSetting<double>("ADFIO2", true, true), // oxygen level, 0.20 -> 1.0
-                new VentilatorSetting<double>("RRHIGHLIMIT", true, true), // detect hyperventilating, absolute value
+                new VentilatorSetting<int>("RRHIGHLIMIT", true, true), // detect hyperventilating, absolute value
             };
         }
     }
